# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This documentation describes MiniJob API",
        "title": "MiniJob API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/nbigot/minijob/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/job/": {
            "post": {
                "description": "Create job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Create job",
                "operationId": "job-create",
                "parameters": [
                    {
                        "description": "Job creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/job.JobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/web.JSONResultCreateJob"
                        }
                    },
                    "400": {
                        "description": "Invalid input - content type, JSON body format or validation errors",
                        "schema": {
                            "$ref": "#/definitions/github_com_nbigot_minijob_web_apierror.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/job/{jobuuid}": {
            "get": {
                "description": "Get a job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a job",
                "operationId": "job-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job UUID",
                        "name": "jobuuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/web.JSONResultGetJob"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_nbigot_minijob_web_apierror.APIError"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nbigot_minijob_web_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Invalid job",
                        "schema": {
                            "$ref": "#/definitions/github_com_nbigot_minijob_web_apierror.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete job",
                "operationId": "job-delete",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/web.JSONResultSuccess"
                        }
                    }
                }
            }
        },
...
        "/api/v1/jobs/old": {
            "get": {
                "description": "Get top N oldest jobs that are not completed, grouped by topic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get oldest incomplete jobs",
                "operationId": "jobs-get-old",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of jobs per topic (default 10, min 1, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum job duration in seconds (default 0, max 1 year)",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/web.JSONResultGetOldJobs"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_nbigot_minijob_web_apierror.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "details": {
                    "description": "application-level error details that best describes the error, for debugging",
                    "type": "string"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "jobUUID": {
                    "description": "job uuid",
                    "type": "string"
                },
                "validationErrors": {
                    "description": "list of errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_nbigot_minijob_web_apierror.ValidationError"
                    }
                }
            }
        },
        "github_com_nbigot_minijob_web_apierror.ValidationError": {
            "type": "object",
            "properties": {
                "failedField": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "job.Job": {
            "type": "object",
            "properties": {
                "debugMode": {
                    "description": "The debug flag of the job (optional)",
                    "type": "boolean"
                },
                "history": {
                    "description": "The list of events of the job",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/job.JobHistoryEvent"
                    }
                },
                "id": {
                    "description": "The unique identifier of a job (required)",
                    "type": "string"
                },
                "lockResources": {
                    "description": "The list of resources to lock (optional)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "The name of the job (optional)",
                    "type": "string"
                },
                "priority": {
                    "description": "The priority of the job (optional)",
                    "type": "integer"
                },
                "properties": {
                    "description": "The job properties (required)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/job.JobProperties"
                        }
                    ]
                },
                "requester": {
                    "description": "The identifier of the job requester (optional)",
                    "type": "string"
                },
                "sessionId": {
                    "description": "The session identifier of the requester (optional)",
                    "type": "string"
                },
                "startAfter": {
                    "description": "Timestamp (in milliseconds) to start the job after",
                    "type": "integer"
                },
                "topic": {
                    "description": "The topic name for which the job has been created (optional)",
                    "type": "string"
                },
                "traceId": {
                    "description": "The trace identifier of the job (optional)",
                    "type": "string"
                },
                "userAgent": {
                    "description": "The user agent (or program name) that make the request (optional)",
                    "type": "string"
                },
                "visibilityTimeout": {
                    "description": "Duration (in seconds) to keep the job hidden from the queue after it is fetched",
                    "type": "integer"
                }
            }
        },
        "job.JobHistoryEvent": {
            "type": "object",
            "properties": {
                "eventType": {
                    "description": "The event type (required)",
                    "type": "string"
                },
                "timestamp": {
                    "description": "The event timestamp (unixmilliseconds) (required)",
                    "type": "integer"
                }
            }
        },
        "job.JobProperties": {
            "type": "object",
            "additionalProperties": true
        },
        "job.JobRequest": {
            "type": "object",
            "required": [
                "properties"
            ],
            "properties": {
                "debugMode": {
                    "description": "DebugMode is the debug flag of the job (optional)",
                    "type": "boolean"
                },
                "delay": {
                    "description": "Delay (in seconds) to wait before starting the job",
                    "type": "integer"
                },
                "lockResources": {
                    "description": "LockResources is the list of resources to lock (optional)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name is the name of the job (optional)",
                    "type": "string",
                    "example": "myJobName01"
                },
                "priority": {
                    "description": "Priority is the priority of the job (optional)",
                    "type": "integer"
                },
                "properties": {
                    "description": "Properties is the job properties (required)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/job.JobProperties"
                        }
                    ]
                },
                "requester": {
                    "description": "Requester is the identifier of the job requester (optional)",
                    "type": "string"
                },
                "sessionId": {
                    "description": "SessionId is the session identifier of the requester (optional)",
                    "type": "string",
                    "example": "6e70464d-85cf-431d-a9c9-a1f9c0ac82dc"
                },
                "startAfter": {
                    "description": "Timestamp (in milliseconds) to start the job after",
                    "type": "integer"
                },
                "topic": {
                    "description": "Topic is the topic name for which the job has been created (optional)",
                    "type": "string",
                    "example": "default"
                },
                "traceId": {
                    "description": "TraceId is the trace identifier of the job (optional)",
                    "type": "string",
                    "example": "e7b46b16-c971-464e-8a47-c791be0ce2ed"
                },
                "userAgent": {
                    "description": "UserAgent is the user agent (or program name) that make the request (optional)",
                    "type": "string",
                    "example": "myJobProducerName"
                }
            }
        },
...
        "web.JSONResultCreateJob": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The result code",
                    "type": "integer",
                    "example": 200
                },
                "job": {
                    "description": "The job",
                    "allOf": [
                        {
                            "$ref": "#/definitions/job.Job"
                        }
                    ]
                },
                "message": {
                    "description": "The result message",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "web.JSONResultGetJob": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The result code",
                    "type": "integer",
                    "example": 200
                },
                "job": {
                    "description": "The job",
                    "allOf": [
                        {
                            "$ref": "#/definitions/job.Job"
                        }
                    ]
                },
                "message": {
                    "description": "The result message",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "web.JSONResultGetOldJobs": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The result code",
                    "type": "integer"
                },
                "message": {
                    "description": "The result message",
                    "type": "string"
                },
                "topics": {
                    "description": "The oldest jobs by topic",
                    "allOf": [
                        {
                            "$ref": "#/definitions/web.TopicOldJobsMap"
                        }
                    ]
                }
            }
        },
...
        "web.JSONResultSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The result code",
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "description": "The result message",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "web.TopOldJobsResult": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "The job duration in milliseconds",
                    "type": "integer"
                },
                "jobuuid": {
                    "description": "The job UUID",
                    "type": "string"
                }
            }
        },
        "web.TopicOldJobsMap": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "$ref": "#/definitions/web.TopOldJobsResult"
                }
            }
        }
    }
}